requirements.txt
appmap>=1.1.0.dev0
langgraph
langsmith
langchain-fireworks
langchain-pinecone
langchain
langchain-openai
langchain-anthropic
langgraph-sdk
langchain-community
tavily-python
tiktoken
IPython
anthropic
python-dotenv
pinecone
fastapi
uvicorn
aiohttp
rich